Test coverage is completed

+++ src/main/java/com/stshysha/app/classwork/lesson4/OddNumber.java
@@ -4,7 +4,7 @@ package com.stshysha.app.classwork.lesson4;
  * Created by new on 21.02.17.
  */
 public class OddNumber {
-    public boolean isOdd (int a) {
+    public boolean isOdd (double a) {
         double b = a/2;
         if (a == ((int)b)*2){
             return false;
